# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fill.me API
  description: |
    Fill.me provides a comprehensive REST API for form creation, management, and analytics. 
    This API enables developers to integrate form functionality into their applications and manage form data programmatically.
    
    ## Authentication
    
    Fill.me API supports multiple authentication methods:
    - **JWT Token**: `Authorization: Bearer YOUR_JWT_TOKEN`
    - **API Key**: `X-API-Key: YOUR_API_KEY`
    - **Basic Auth**: `Authorization: Basic base64(email:password)`
    
    ## Rate Limiting
    
    Rate limits vary by subscription plan:
    - Free: 100 requests/hour
    - Starter: 500 requests/hour
    - Professional: 2,000 requests/hour
    - Enterprise: 10,000 requests/hour
    
  version: 1.3.0
  contact:
    name: Fill.me Support
    email: support@fill.me
    url: https://fill.me/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.fill.me
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - BasicAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    Form:
      type: object
      properties:
        id:
          type: string
          example: "form_123"
        title:
          type: string
          example: "Contact Form"
        description:
          type: string
          example: "Get in touch with us"
        status:
          type: string
          enum: [draft, active, archived]
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
        settings:
          type: object
          properties:
            allowMultiple:
              type: boolean
            requireAuth:
              type: boolean
            collectIP:
              type: boolean
        customization:
          type: object
          properties:
            theme:
              type: string
            colors:
              type: object
              properties:
                primary:
                  type: string
                background:
                  type: string
        analytics:
          type: object
          properties:
            views:
              type: integer
            submissions:
              type: integer
            conversionRate:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FormField:
      type: object
      properties:
        id:
          type: string
          example: "field_1"
        type:
          type: string
          enum: [text, email, number, textarea, dropdown, radio, checkbox, date, file]
        label:
          type: string
          example: "Full Name"
        placeholder:
          type: string
          example: "Enter your name"
        required:
          type: boolean
        order:
          type: integer
        options:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              label:
                type: string
        validation:
          type: object
          properties:
            minLength:
              type: integer
            maxLength:
              type: integer
            pattern:
              type: string
        conditional:
          type: object
          properties:
            enabled:
              type: boolean
            conditions:
              type: array
              items:
                type: object
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "The provided data is invalid"
        details:
          type: object
...
paths:
...
  /forms/{formId}:
    get:
      tags: [Forms]
      summary: Get form by ID
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Form retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Form'
        '404':
          description: Form not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Forms]
      summary: Update form
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/FormField'
                settings:
                  type: object
                customization:
                  type: object
      responses:
        '200':
          description: Form updated successfully

    delete:
      tags: [Forms]
      summary: Delete form
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Form deleted successfully
...
  /forms/{formId}/analytics:
    get:
      tags: [Analytics]
      summary: Get form analytics
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - name: timeRange
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
        - name: metrics
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      overview:
                        type: object
                        properties:
                          totalViews:
                            type: integer
                          totalSubmissions:
                            type: integer
                          conversionRate:
                            type: number
                          averageTime:
                            type: integer
                      timeline:
                        type: array
                        items:
                          type: object
                      devices:
                        type: object
                      locations:
                        type: array

  # Templates
...